Add shop entrypoint with basic functionality
